refactor: move TagList to it's own file (see below) o move TagList class from views.py to new TagList.py file o rename test_views.py to test_TagList.py
